requirements.txt
black==19.10b0
connexion==2.6.0
flask-mongoengine==0.9.5
pip-chill==0.1.8
swagger-ui-bundle==0.0.6
